+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-Improve documentation
+- Improve documentation
+- Better handling of full paths with `:root` value
 
 ## 0.1.0